Set a fallback value when reading CASTEP XC functional
CASTEP has recently reformatted LDA functionals, and can also
construct functionals not in this table using libxc, hybrid DFT...

This attempt to reconstruct the params from the .castep file _cannot_
work for all cases across CASTEP versions without an unreasonable
maintenance overhead; it is much more sensible to read the .param
file.

To avoid KeyError when reading such calcs, here we use the whole
string as a fallback. This will give a fairly meaningful calculation
dump, but not a valid CASTEP input - a restart without using the
.param file will quickly and noisily fail.

Alternatively we could fallback to None, and then if you run CASTEP it
will use LDA. That seems the worse outcome as it would lead to
"quietly incorrect" results on a restart.